requirements.txt
python-docx==1.1.2  # For .docx parsing/editing
langchain==0.3.0  # For RAG
langchain-community==0.3.0
langchain-groq==0.2.0  # Groq integration
sentence-transformers==3.1.1  # Embeddings
faiss-cpu==1.9.0  # Vector store (use faiss-gpu if NVIDIA GPU available)
pypdf2==3.0.1  # PDF parsing
docx2txt==0.8  # .docx text extraction
streamlit==1.39.0  # UI